# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dara Meetings API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TokenRequest {
  roomName: string;
  participantName: string;
  isHost?: boolean;
}

export interface TokenResponse {
  token: string;
  wsUrl: string;
}

export interface Meeting {
  id: string;
  roomName: string;
  title: string;
  hostName: string;
  createdAt: string;
  participantCount: number;
  isActive: boolean;
  maxParticipants?: number;
}

export interface MeetingInput {
  title: string;
  hostName: string;
  maxParticipants?: number;
}

export interface MeetingStats {
  activeMeetings: number;
  totalParticipants: number;
  maxMeetingsAllowed: number;
  meetingsList?: Meeting[];
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface AiChatInput {
  message: string;
  roomId: string;
  conversationHistory?: ChatMessage[];
}

export interface ChatReply {
  reply: string;
}

export interface SummaryInput {
  transcript: string;
  roomId: string;
  bookMode?: boolean;
}

export interface SummaryResult {
  summary: string;
}