# lib/api-zod/src/generated/types/aiChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dara Meetings API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface AiChatInput {
  message: string;
  roomId: string;
  conversationHistory?: ChatMessage[];
}