# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dara Meetings API
 * OpenAPI spec version: 0.1.0
 */

export interface Meeting {
  id: string;
  roomName: string;
  title: string;
  hostName: string;
  createdAt: string;
  participantCount: number;
  isActive: boolean;
  maxParticipants?: number;
}