# lib/api-zod/src/generated/types/meetingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dara Meetings API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingInput {
  title: string;
  hostName: string;
  maxParticipants?: number;
}