# lib/api-zod/src/generated/types/meetingStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dara Meetings API
 * OpenAPI spec version: 0.1.0
 */
import type { Meeting } from './meeting';

export interface MeetingStats {
  activeMeetings: number;
  totalParticipants: number;
  maxMeetingsAllowed: number;
  meetingsList?: Meeting[];
}